refactor: move DbAuthentication usecase to correct folder
