+++ src/data/usecases/authentication/db-authentication.spec.ts
@@ -1,6 +1,6 @@
-import { AuthenticationModel } from '../../../../domain/usecases/authentication'
-import { LoadAccountByEmailRepository } from '../../../protocols/load-account-by-email-repository'
-import { AccountModel } from '../db-add-account-protocols'
+import { AuthenticationModel } from '../../../domain/usecases/authentication'
+import { LoadAccountByEmailRepository } from '../../protocols/load-account-by-email-repository'
+import { AccountModel } from '../add-account/db-add-account-protocols'
 import { DbAuthentication } from './db-authentication'
 
 const makeFakeAuthentication = (): AuthenticationModel => ({
