+++ src/data/usecases/authentication/db-authentication.ts
@@ -1,8 +1,8 @@
 import {
   Authentication,
   AuthenticationModel,
-} from '../../../../domain/usecases/authentication'
-import { LoadAccountByEmailRepository } from '../../../protocols/load-account-by-email-repository'
+} from '../../../domain/usecases/authentication'
+import { LoadAccountByEmailRepository } from '../../protocols/load-account-by-email-repository'
 
 export class DbAuthentication implements Authentication {
   constructor(
